hermioneFiles/src/Luna/requirements.txt
Flask==0.12.2
Werkzeug<0.13.0,>=0.12.0
pymongo==3.6.0
requests==2.18.4
flask_cors==2.1.2
gunicorn==19.3.0
google-api-python-client
google-cloud-bigquery
google-cloud
requests-toolbelt